# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "하드웨어 기반 라이선스 관리 서버",
        "title": "Studio License Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/license/activate": {
            "post": {
                "description": "라이선스 키를 사용하여 디바이스를 활성화합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "클라이언트 - 라이선스"
                ],
                "summary": "라이선스 활성화",
                "parameters": [
                    {
                        "description": "활성화 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ActivateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "이미 활성화됨",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "201": {
                        "description": "활성화 성공",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "403": {
                        "description": "라이선스 비활성/만료/기기수 초과",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "라이선스 없음",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/license/validate": {
            "post": {
                "description": "앱 실행 시 라이선스를 검증합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "클라이언트 - 라이선스"
                ],
                "summary": "라이선스 검증",
                "parameters": [
                    {
                        "description": "검증 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "검증 성공",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "403": {
                        "description": "라이선스 비활성/만료 또는 디바이스 미등록",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "라이선스 없음",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "success, error",
                    "type": "string"
                }
            }
        },
        "models.ActivateRequest": {
            "type": "object",
            "required": [
                "device_info",
                "license_key"
            ],
            "properties": {
                "device_info": {
                    "$ref": "#/definitions/models.DeviceInfo"
                },
                "license_key": {
                    "type": "string"
                }
            }
        },
...
        "models.DeviceInfo": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "cpu_id": {
                    "type": "string"
                },
                "disk_serial": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "mac_address": {
                    "type": "string"
                },
                "machine_id": {
                    "type": "string"
                },
                "motherboard_sn": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "os_version": {
                    "type": "string"
                }
            }
        },
...
        "models.ValidateRequest": {
            "type": "object",
            "required": [
                "device_info",
                "license_key"
            ],
            "properties": {
                "device_info": {
                    "$ref": "#/definitions/models.DeviceInfo"
                },
                "license_key": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 토큰을 입력하세요. 형식: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}